Refactor CLI subcommands into single invoke pattern and convert handlers to static APIs
