--- unity-package/Editor/Handlers/HierarchyHandler.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: e581ca9fefc614d369f259c58ff20511
