--- unity-package/Editor/Handlers/InputHandler.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 1fa8aab5921ff4ff29d904b4e4753c70
