--- unity-package/Editor/Handlers/PlayModeHandler.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 98159dab41cf2484b8077cc7f6241b51
