--- unity-package/Editor/Handlers/SceneHandler.cs
@@ -1,77 +1,0 @@
-using System;
-using System.Collections.Generic;
-using System.Text;
-using System.Threading.Tasks;
-using UnityEditor;
-using UnityEditor.SceneManagement;
-using UnityEngine.SceneManagement;
-
-namespace UnityAgenticTools.Server
-{
-    public class SceneHandler : IRequestHandler
-    {
-        public string MethodPrefix => "editor.scene.";
-
-        public async Task<object> HandleAsync(string method, Dictionary<string, object> parameters)
-        {
-            var action = method.Substring(MethodPrefix.Length);
-
-            switch (action)
-            {
-                case "save":
-                    return await EditorWebSocketServer.RunOnMainThread(() =>
-                    {
-                        var saved = EditorSceneManager.SaveOpenScenes();
-                        if (saved)
-                        {
-                            return new Dictionary<string, object>
-                            {
-                                { "success", true },
-                                { "message", "All open scenes saved" }
-                            };
-                        }
-
-                        var details = new StringBuilder("Failed to save scenes.");
-                        for (int i = 0; i < SceneManager.sceneCount; i++)
-                        {
-                            var scene = SceneManager.GetSceneAt(i);
-                            var scenePath = string.IsNullOrEmpty(scene.path) ? "(unsaved)" : scene.path;
-                            details.Append($" [{scene.name}: path={scenePath}, dirty={scene.isDirty}, loaded={scene.isLoaded}]");
-                        }
-                        throw new InvalidOperationException(details.ToString());
-                    });
-
-                case "open":
-                    return await EditorWebSocketServer.RunOnMainThread(() =>
-                    {
-                        if (!parameters.TryGetValue("path", out var pathObj) || !(pathObj is string scenePath))
-                        {
-                            throw new ArgumentException("Missing required parameter: path");
-                        }
-
-                        var additive = false;
-                        if (parameters.TryGetValue("additive", out var additiveObj))
-                        {
-                            additive = additiveObj is bool b ? b : false;
-                        }
-
-                        var mode = additive
-                            ? OpenSceneMode.Additive
-                            : OpenSceneMode.Single;
-
-                        var scene = EditorSceneManager.OpenScene(scenePath, mode);
-                        return new Dictionary<string, object>
-                        {
-                            { "success", scene.IsValid() },
-                            { "name", scene.name },
-                            { "path", scene.path },
-                            { "buildIndex", scene.buildIndex }
-                        };
-                    });
-
-                default:
-                    throw new InvalidOperationException($"Unknown scene action: {action}");
-            }
-        }
-    }
-}
