--- unity-package/Editor/Handlers/SceneHandler.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 1d3c5b616a7214a73aa83553cc342add
