--- unity-package/Editor/Handlers/ScreenshotHandler.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 990c372b5f3e2479fa72da3c1987075d
