--- unity-package/Editor/Handlers/TestRunnerHandler.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 5a9b9b6c40a774a59b8fc7817f3ff39b
