--- unity-package/Editor/Handlers/UIHandler.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: fc49f67b861e147c1bf090c76713c077
